last/back/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
mysql-connector-python
pandas
scikit-learn
numpy
python-dotenv